refactor(usecases): ユースケース層の導入によるDDDの深化とコントローラーの非同期化
- コントローラーからの直接のデータベースアクセス（IScrapingContextFactory）を排除
- Application/Use Cases層として4つのユースケースクラスを新規実装
- Startup.csにて各ユースケースをScopedサービスとしてDIコンテナへ登録
- BukkenControllerおよびFileControllerのアクションメソッドをasync/awaitによる非同期処理へリファクタリング
- CancellationTokenを伝播させ、クエリのキャンセル処理に対応
